requirements.txt
altgraph==0.17.5
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
et_xmlfile==2.0.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lapx==0.9.4
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.90
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pefile==2024.8.26
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
psutil==7.2.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
setuptools==80.10.2
six==1.17.0
sympy==1.14.0
tkinterdnd2==0.4.3
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.9
ultralytics-thop==2.0.18
urllib3==2.6.3
yt-dlp==2026.1.31
